Rename back to kubespawner
The two 's' in kubernetesspawner was an eyesore, and kubespawner
has enough precursors for it to be ok.

This reverts commit e4414357438408d0b0704323f0718f71a1955b3f.
