+++ jupyterhub_config.py
@@ -1,7 +1,7 @@
 import os
 
 
-c.JupyterHub.spawner_class = 'kubernetesspawner.KubernetesSpawner'
+c.JupyterHub.spawner_class = 'kubespawner.KubeSpawner'
 
 c.JupyterHub.ip = '0.0.0.0'
 c.JupyterHub.hub_ip = '0.0.0.0'
@@ -11,13 +11,13 @@ c.JupyterHub.hub_ip = '0.0.0.0'
 c.JupyterHub.cleanup_servers = False
 
 # First pulls can be really slow, so let's give it a big timeout
-c.KubernetesSpawner.start_timeout = 60 * 5
+c.KubeSpawner.start_timeout = 60 * 5 
 
 # Our simplest user image! Optimized to just... start, and be small!
-c.KubernetesSpawner.singleuser_image_spec = 'yuvipanda/simple-singleuser:v1'
+c.KubeSpawner.singleuser_image_spec = 'yuvipanda/simple-singleuser:v1'
 
 # The spawned containers need to be able to talk to the hub through the proxy!
-c.KubernetesSpawner.hub_connect_ip = os.ENVIRON['HUB_CONNECT_IP']
+c.KubeSpawner.hub_connect_ip = os.ENVIRON['HUB_CONNECT_IP']
 
 # Do not use any authentication at all - any username / password will work.
 c.JupyterHub.authenticator_class = 'dummyauthenticator.DummyAuthenticator'
