+++ setup.py
@@ -1,7 +1,7 @@
-from setuptools import setup, find_packages
+from setuptools import setup
 
 setup(
-    name='jupyterhub-kubernetesspawner',
+    name='jupyterhub-kubespawner',
     version='0.1',
     install_requires=[
         'jupyterhub',
@@ -12,7 +12,7 @@ setup(
     description='JupyterHub Spawner targetting Kubernetes',
     url='http://github.com/yuvipanda/jupyterhub-kubernetes-spawner',
     author='Yuvi Panda',
-    author_email='yuvipanda@gmail.com',
+    author_email='yuvipanda@riseup.net',
     license='BSD',
-    packages=find_packages(),
+    packages=['kubespawner'],
 )
